requirements.txt
PyMuPDF==1.23.8
pdf2image==1.17.0
ultralytics==8.1.2
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0 